+++ modulo5/IWFS-LaBook/requests.rest
@@ -1,4 +1,5 @@
 GET http://localhost:3003/user
+Authorization: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjExOGNlZGE5LTZmM2ItNGY1MS1iOTlhLWEyZDJmMTY1MTVkZCIsInJvbGUiOiJBRE1JTiIsImlhdCI6MTY0NzAwNzgzOCwiZXhwIjoxNjQ3MDA4NDM4fQ.knX81Umr30FzzX6MEMbVWvia09AtEpUH8vMKrymwXI8"
 
 
 ###
